third_party/WebKit/LayoutTests/inspector/console/console-trace-expected.txt

CONSOLE MESSAGE: line 9: console.trace
Tests that console.trace dumps stack trace with source URLs and line numbers.

console-trace.html:9 console.trace
b @ console-trace.html:9
a @ console-trace.html:14
